made some changes
