+++ Oras/Constants/DockerMediaTypes.cs
@@ -1,6 +1,6 @@
 ﻿namespace Oras.Constants
 {
-    public class DockerMediaTypes
+    public static class DockerMediaTypes
     {
         // Docker media types
         public const string Config = "application/vnd.docker.container.image.v1+json";
