+++ Oras/Interfaces/IStorage.cs
@@ -6,13 +6,13 @@ using System.Threading.Tasks;
 namespace Oras.Interfaces
 {
     /// <summary>
-    /// IStorage represents a _content-addressable _storage (CAS) where contents are accessed via Descriptors.
+    /// IStorage represents a content-addressable storage (CAS) where contents are accessed via Descriptors.
     /// The storage is designed to handle blobs of large sizes.
     /// </summary>
     public interface IStorage : IReadOnlyStorage
     {
         /// <summary>
-        /// PushAsync pushes the _content, matching the expected descriptor.
+        /// PushAsync pushes the content, matching the expected descriptor.
         /// </summary>
         /// <param name="expected"></param>
         /// <param name="content"></param>
